Handle invalid or missing author details in commit
Handles the case where there is an author for a commit but its not a valid/complete author. I.E. when there is no html_url for the author.

This can happen if the commit was made on a system that has not set the author email in the git config properly.

In this case it will not add the invalid author to the contributors list and will not improperly increment an author commit count.